requirements.txt
google-cloud-bigquery==3.25.0
functions-framework==3.0.0
